+++ flask_xxljob/__init__.py
@@ -34,7 +34,7 @@ execute the actual tasks.
 
 from __future__ import annotations
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 
 from .client import AdminCallResult, CallResult
 from .exceptions import (